TypeScript validation during Metro bundling (#392)
* Update typescript-service:
- Separate Project methods into "find" and "open", allowing caller to decide on the appropriate error-handling logic.
- Add filtering when validating Project files. Only include ts[x] unless checkJs is enabled.
- Make cache methods return a true/false result, moving error-handling up and out to the caller (Project).
- Stop converting to lowercase during path normalization. This confuses typescript.

* Add typescript validation to CLI rnx-bundle call.

* Don't check JS files when bundling test-app, as this would expose all of react-native JS files. They aren't typesafe and fail the check.

* Fix clean command (build) to clear out the lage cache, ensuring that everything is rebuilt and retested.

* Change files

* change file

* PR feedback

+++ packages/typescript-service/src/cache.ts
@@ -9,23 +9,41 @@ export class ProjectFileCache {
     fileNames.forEach((fileName) => this.add(fileName));
   }
 
-  add(fileName: string): void {
+  has(fileName: string): boolean {
     const normalized = normalizePath(fileName);
-    this.files.set(normalized, new VersionedSnapshot(normalized));
+    return this.files.has(normalized);
   }
 
-  update(fileName: string, snapshot?: ts.IScriptSnapshot): void {
+  add(fileName: string): boolean {
     const normalized = normalizePath(fileName);
-    const file = this.files.get(normalized);
-    if (!file) {
-      throw new Error(`Cannot update unknown project file ${fileName}`);
+    if (this.files.has(normalized)) {
+      return false;
     }
-    file.update(snapshot);
+    this.files.set(normalized, new VersionedSnapshot(normalized));
+    return true;
   }
 
-  delete(fileName: string): void {
+  update(fileName: string, snapshot?: ts.IScriptSnapshot): boolean {
     const normalized = normalizePath(fileName);
+    if (!this.files.has(normalized)) {
+      return false;
+    }
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- Map has the element, and we only store non-null objects
+    this.files.get(normalized)!.update(snapshot);
+    return true;
+  }
+
+  delete(fileName: string): boolean {
+    const normalized = normalizePath(fileName);
+    if (!this.files.has(normalized)) {
+      return false;
+    }
     this.files.delete(normalized);
+    return true;
+  }
+
+  deleteAll(): void {
+    this.files.clear();
   }
 
   getFileNames(): string[] {
@@ -56,6 +74,7 @@ export class ExternalFileCache {
       this.files.set(normalized, file);
     }
 
-    return file.getSnapshot();
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion -- Map has the element, and we only store non-null objects
+    return this.files.get(normalized)!.getSnapshot();
   }
 }
